requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==2.0.0
qrcode==7.4.2
reportlab==4.0.4
Werkzeug==2.3.0
gunicorn==21.2.0